+++ app/Mail/ConfirmAppEmail.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace App\Mail;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
+use Illuminate\Mail\Mailables\Content;
+use Illuminate\Mail\Mailables\Envelope;
+use Illuminate\Queue\SerializesModels;
+
+class ConfirmAppEmail extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    public $verify;
+    public $appUrl;
+
+    /**
+     * Create a new message instance.
+     *
+     * @return void
+     */
+    public function __construct($verify, $url)
+    {
+        $this->verify = $verify;
+        $this->appUrl = $url;
+    }
+
+    /**
+     * Get the message envelope.
+     *
+     * @return \Illuminate\Mail\Mailables\Envelope
+     */
+    public function envelope()
+    {
+        return new Envelope(
+            subject: 'Complete Account Registration',
+        );
+    }
+
+    /**
+     * Get the message content definition.
+     *
+     * @return \Illuminate\Mail\Mailables\Content
+     */
+    public function content()
+    {
+        return new Content(
+            markdown: 'emails.confirm_app_email',
+            with: [
+                'verify' => $this->verify,
+                'appUrl' => $this->appUrl
+            ],
+        );
+    }
+
+    /**
+     * Get the attachments for the message.
+     *
+     * @return array
+     */
+    public function attachments()
+    {
+        return [];
+    }
+}